Fix non-empty container render error
We need to clear non-empty containers on their first render
or the DOM child indexing will be forever wrong and updates
will crash.

Also added a (hopefully never needed) diagnostic error message
to the console if update() encounters a dom_elem that doesn't
exist again. I presume this would happen if manual changes are
made to the DOM outside RetroV.